Added new spiritual identifier item with magnifying glass texture, no major features yet but going to be used to identify cultivators
