+++ src/main/resources/assets/daoism/lang/en_us.lang
@@ -1,1 +1,2 @@
 item.food_aquapill.name=Aqua Pill
+item.misc_magnifying_glass.name=Spiritual Identifier
