+++ src/main/resources/assets/daoism/models/item/misc_magnifying_glass.json
@@ -1,0 +1,11 @@
+{
+  "parent": "item/generated",
+  "textures": {
+    "layer0": "daoism:item/misc_magnifying_glass"
+  },
+    "firstperson": {
+      "rotation": [0,-135,25],
+      "translation": [0,4,2],
+      "scale": [1.7,1.7,1.7]
+    }
+  }
